requirements.txt
alembic==1.4.3
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
coverage==5.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.0
itsdangerous==1.1.0
Jinja2==2.11.2
kaggle==1.5.8
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
